Add onAction middleware to root store per MST warning. Be sure to reset sesssion count too
